Use base testclass for workflow config test
